dashboards/pyproject.toml
[tool.poetry]
name = "tachyus-dashboards"
version = "0.1.0"
description = "Tachyus Dashboards"
authors = ["Santiago Cuervo <santiago.cuervo@tachyus.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "tachyus_dashboards"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
numpy = "^1.25.0"
dash = "^2.11.1"
plotly = "^5.15.0"
gunicorn = "^20.1.0"
dash-bootstrap-components = "^1.4.1"
dash-daq = "^0.5.0"
httpx = "^0.24.1"
dash-ag-grid = "^2.2.0"
pypika = "^0.48.9"
python-dotenv = "^1.0.0"
palettable = "^3.3.3"
matplotlib = "^3.7.2"

[tool.poetry.group.dev.dependencies]
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^1.4.41"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
